+++ src/main/java/lotto/Application.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.service.LottoMachine;
+import lotto.controller.LottoMachine;
 
 public class Application {
     public static void main(String[] args) {
